Crate delf

DeLF

A DelF inspired deletion framework in Rust.

The goal of DelF is to ensure that when an item is deleted (such as a user in a social network), all of their related data is also deleted (photos, friendships, posts, etc.). DelF achieves this by defining the backing storage as a graph then using methods on that graph to delete the object/edge. DelF will then traverse the graph ensuring deletion of all related objects/edges.

DelF Data Definition Language (DDL)

The DelF DDL enables the functionality within the delf crate. A DelF schema is a YAML file containing definitions of objects and their outbound edges. The DDL is as follows:

--- # the triple line indicates a new document
# root key of the document
object_type:

  # The name of the object, this is a unique identifier for this object, such as a table name
  # in a relational database
  name: MyObjectName

  # The storage is the name of a storage in the config file
  storage: my_storage

  # The deletion type of the object - see delf::graph::object::DeleteType for details
  deletion: by_any | by_x_only | short_ttl | directly | directly_only | not_deleted

  # The field or attribute of the object that holds a unique identifier for each instance
  id: 123

  # If the deletion specified is `by_x_only`, provide an array of inbound edge names that can
  # delete this object
  x:
    - my_edge
    - my_other_edge

  # If the deletion specified is `short_ttl`, specify the field or attribute holding the time
  # an instance of the object should be deleted
  time_field: my_delete_time

  # Definitions of outbound edges from this object to other objects.  If no edges exists, can
  # pass an empty array (`[]`)
  edge_types:

    # A unique name for the edge that can be referenced from other definitions and to directly
    # delete an edge directly
    - name: edge_to_somewhere

      # The deletion type of the edge - see delf::graph::edge::DeleteType for details
      deletion: shallow | deep | refcount

      # An inverse edge may be specified, typically with a deletion type of `shallow`, to
      # delete an edge in the opposite direction of this edge when this edge is deleted
      inverse: edge_to_here

      # Fields describing the object this edge points to (from the edge's point of view)
      to:

        # The name of the object_type this edge points to
        object_type: MyOtherObject

        # The name of the field/attribute the id of this object references on the to object
        field: my_object_name_id

        # Optionally, if there's an intermediate mapping table between the two objects, (e.g.
        # user -> user_photos -> photos ) - more common in relational databases
        mapping_table: this_2_other

Configuration

A configuration YAML file is also required to use delf. It contains information on the storage infrastructure delf will be acting on (such as how to connect to a database). The format is as follows:

# A list of the storage infrastruture delf is acting on
storages:

  # The name of the storage as referenced in the DDL
  - name: my_storage

    # The name of the plugin (see delf::storage::DelfStorageConnection) this storage uses
    plugin: diesel (other plugins may be implemented in the future)

    # A connection string for the storage
    url: mysql://username:password@localhost:port/database_name

Command Line Interface (CLI)

The DelF CLI provides two commands: validate and run.

Validate

The validate command checks that all of the objects defined in the schema exist as described in the storage and also that all objects in the delf graph are reachable by traversing the graph starting from a directly deletable object then following edges that can delete the connecting object. This ensures that all objects are deletable (or excepted from deletion by using the not_deleted type).

Example

delf -s path/to/schema.yaml -c path/to/config.yaml validate

The executable will print a report of its findings.

Run

The run command starts the api running locally and starts a thread which checks short_ttl deleted objects for deletable instances every 30 seconds.

Example

delf -s path/to/schema.yaml -c path/to/config.yaml run

The executable will run until terminated.

Modules

api
The api module contains routes to dispatch deletes on objects and edges.

graph
The graph module contains the core structures to run the deletion algorithms.

storage
The storage module contains plugins for storage-specific (e.g. mysql) deletion implementations.

Structs

DelfYamls
Convenience structure for passing around parsed yaml files

Functions

check_short_ttl_loop
Spawn a thread that checks short time to live objects every 30 seconds to evaluate for deletion

init_api
Initialize the rocket api and return the struct. Run the launch method on the returned value to start the api.

read_files
Read in the schema and config yaml files to construct the delf graph.

read_yamls
Read in the schema and config yaml strings to construct the delf graph.