Module delf::api

The api module contains routes to dispatch deletes on objects and edges.

The two routes the api uses are for deletion of objects and edges:

object: DELETE /object/<object_type>/<id> edge: DELETE /edge/<edge_type>/<from_id>/<to_id>

Example

object: DELETE /object/users/123 edge: DELETE /edge/user_photo/123/my_photo456

Statics

static_rocket_route_info_for_delete_edge
Rocket code generated static route info.

static_rocket_route_info_for_delete_object
Rocket code generated static route info.

Functions

delete_edge
DELETE API endpoint to delete an edge of the given type and the IDs of the objects it is connecting.

delete_object
DELETE API endpoint to delete an object of the given type with specified ID.

rocket_route_fn_delete_edge
Rocket code generated wrapping route function.

rocket_route_fn_delete_object
Rocket code generated wrapping route function.