Trait delf::storage::DelfStorageConnection

pub trait DelfStorageConnection: Debug {
    pub fn connect(database_url: &str) -> Self
    where
        Self: Sized;

    pub fn get_object_ids(
        &self, 
        from_id: &String, 
        from_id_type: &String, 
        edge_field: &String, 
        table: &String, 
        id_field: &String, 
        id_type: &String
    ) -> Vec<String>;

    pub fn get_object_ids_by_time(
        &self, 
        table: &String, 
        time_field: &String, 
        id_field: &String, 
        id_type: &String
    ) -> Vec<String>;

    pub fn delete_edge(
        &self, 
        to: &DelfObject, 
        from_id: &String, 
        to_id: Option<&String>, 
        edge: &DelfEdge
    ) -> bool;

    pub fn delete_object(&self, obj: &DelfObject, id: &String) -> bool;

    pub fn validate_edge(&self, edge: &DelfEdge) -> Result<(), String>;

    pub fn validate_object(&self, obj: &DelfObject) -> Result<(), String>;

    pub fn has_edge(
        &self, 
        obj: &DelfObject, 
        id: &String, 
        edge: &DelfEdge
    ) -> bool;
}

Trait defining the api for a DelF storage connection.

Required methods

pub fn connect(database_url: &str) -> Self where
Self: Sized,

Connect to the storage via a connection string

pub fn get_object_ids(
&self,
from_id: &String,
from_id_type: &String,
edge_field: &String,
table: &String,
id_field: &String,
id_type: &String
) -> Vec<String>

Get the object ids of the object an edge points to.

pub fn get_object_ids_by_time(
&self,
table: &String,
time_field: &String,
id_field: &String,
id_type: &String
) -> Vec<String>

Get the object ids that have a time_field with a value before now.

pub fn delete_edge(
&self,
to: &DelfObject,
from_id: &String,
to_id: Option<&String>,
edge: &DelfEdge
) -> bool

Delete an edge instance.

pub fn delete_object(&self, obj: &DelfObject, id: &String) -> bool

Delete an object instance.

pub fn validate_edge(&self, edge: &DelfEdge) -> Result<(), String>

Validate the edge exists in the storage as described in the struct

pub fn validate_object(&self, obj: &DelfObject) -> Result<(), String>

Validate the object exists in the storage as described in the struct

pub fn has_edge(&self, obj: &DelfObject, id: &String, edge: &DelfEdge) -> bool

Check if an inbound edge exists for a given object instance.

Loading content...